Move code out of package init
