+++ pyang/plugins/check_update.py
@@ -8,13 +8,14 @@ import sys
 import os
 import io
 
-import pyang
+from pyang import repository
 from pyang import plugin
 from pyang import statements
 from pyang import error
 from pyang import util
 from pyang import types
 from pyang.error import err_add
+
 
 def pyang_plugin_init():
     plugin.register_plugin(CheckUpdatePlugin())
@@ -140,8 +141,8 @@ def check_update(ctx, newmod):
     if olddir == '':
         olddir = '.'
     oldpath += os.pathsep + olddir
-    oldrepo = pyang.FileRepository(oldpath, use_env=False)
-    oldctx = pyang.Context(oldrepo)
+    oldrepo = repository.FileRepository(oldpath, use_env=False)
+    oldctx = repository.Context(oldrepo)
     oldctx.opts = ctx.opts
     oldctx.lax_xpath_checks = ctx.lax_xpath_checks
     oldctx.lax_quote_checks = ctx.lax_quote_checks
