+++ test/test_issues/test_i225/test_prefix_deviation.py
@@ -13,7 +13,7 @@ try:
 except ImportError:
     import pip._internal.locations as locations
 
-from pyang import Context, FileRepository
+from pyang.repository import Context, FileRepository
 
 EXISTING_MODULE = 'ietf-yang-types'
